minimize listen.py further
